% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwriteXLSX.R
\name{cellDate}
\alias{cellDate}
\title{Date cell}
\usage{
cellDate(col, row, date, colspan = NULL, rowspan = NULL, comment = NULL,
  commentAuthor = NULL, fontname = NULL, bold = NULL, color = NULL,
  fill = NULL, dateFormat = "yyyy-mm-dd;@", ...)
}
\arguments{
\item{col}{integer, the column index}

\item{row}{integer, the row index}

\item{date}{an input which will be parsed to a date through the
\code{\link[anytime]{anydate}} function}

\item{colspan}{integer, number of columns the cell spans}

\item{rowspan}{integer, number of rows the cell spans}

\item{comment}{cell comment, character}

\item{commentAuthor}{author of the comment, character}

\item{fontname}{font name; see \code{\link{cell}}}

\item{bold}{set \code{TRUE} for bold font}

\item{color}{color name; see \code{\link{cell}}}

\item{fill}{color name; see \code{\link{cell}}}

\item{dateFormat}{the date format}

\item{...}{arguments passed to \code{\link[anytime]{anydate}}}
}
\value{
A named list representing a cell.
}
\description{
Creates a cell for a date.
}
\details{
A date in Excel is stored as an integer: the number of days between
this date and a certain origin. The \code{cellDate} function calculates this
integer and then passes it to the function \code{\link{cell}} with a date format.
}
\examples{
A1 <- cellDate(1, 1, "2017-06-14")
unlist(A1)
A1 <- cellDate(1, 1, as.Date(42900, origin="1899-12-30"))
unlist(A1)
A1 <- cellDate(1, 1, "2017/06/14")
unlist(A1)
A1 <- cellDate(1, 1, "20170614")
unlist(A1)
A1 <- cellDate(1, 1, "06-14-2017")
unlist(A1)
A1 <- cellDate(1, 1, "2017-06-14 04:05:06")
unlist(A1)
A1 <- cellDate(1, 1, "2017-06-14 04:05:06", tz="America/Los_Angeles")
unlist(A1)
A1 <- cellDate(1, 1, Sys.Date())
unlist(A1)
A1 <- cellDate(1, 1, "x")
unlist(A1)
}
