% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwriteXLSX.R
\name{createSheet}
\alias{createSheet}
\title{Create a sheet}
\usage{
createSheet(dat, sheetname)
}
\arguments{
\item{dat}{appropriate data input (see examples and vignette)}

\item{sheetname}{name of the sheet}
}
\value{
A named list containing one element: the list of cells.
}
\description{
Create a sheet from an appropriate data input.
}
\examples{
# Create a sheet from a dataframe
sheet <- createSheet(mtcars[1:2, 1:2], "Sheet1")
# write to xlsx.xlsx:
\donttest{hwriteXLSX(file="xlsx.xlsx", worksheet=sheet)}
# Create a sheet from a list of lists
dat <-
  list(
    A = list(1, 2, 3),
    B = list(NULL, "a", 1000),
    C = list(NA, "b")
  )
attr(dat$B[[2]], "color") <- "red"
attr(dat$B[[2]], "comment") <- "this cell is red"
attr(dat$C[[1]], "comment") <- "this cell is empty"
sheet <- createSheet(dat, "Sheet1")
\donttest{hwriteXLSX("xlsx.xlsx", sheet)}
}
\seealso{
\code{\link{hwriteXLSX}}
}
