% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{guessDateFormat}
\alias{guessDateFormat}
\title{Guess date format}
\usage{
guessDateFormat(x, returnDates = FALSE, tzone = "", EN = TRUE,
  messages = FALSE)
}
\arguments{
\item{x}{a vector (not a factor)}

\item{returnDates}{logical, whether to return the dates}

\item{tzone}{time zone}

\item{EN}{logical; if \code{TRUE}, looks for the format "month-day-year", othewise "day-month-year"}

\item{messages}{logical, whether to print some messages}
}
\value{
\itemize{
     \item If \code{returnDates=FALSE}, the date format if guessed, \code{FALSE} if no date format is guessed.
     \item If \code{returnDates=TRUE}, the date if the date format if guessed, \code{FALSE} if no date format is guessed.
}
}
\description{
Guess the date format.
}
\examples{
guessDateFormat("2017-03-04")
guessDateFormat("2017-03-04", returnDates = TRUE)
guessDateFormat(c("2017-03-04", "2017-04-04"))
guessDateFormat("04-03-2017")
guessDateFormat(c("04-03-2017", "2017-04-04"))
guessDateFormat("13-03-2017")
guessDateFormat("13-03-2017", EN=FALSE)
}
\author{
Cole Beck.
}
