% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{is_date}
\alias{is_date}
\title{Check whether has a date-like class}
\usage{
is_date(x)
}
\arguments{
\item{x}{vector or factor}
}
\value{
A logical value, indicating whether \code{x} has a date-like class.
}
\description{
Check whether a vector has a date-like class.
}
\examples{
x <- Sys.Date()
is_date(x)
x <- strftime(c("2016-06-21", "2016-11-22"), usetz=TRUE, tz="UTC")
is_date(x)
x <- as.POSIXct(x, tz="UTC")
is_date(x)
}
\seealso{
\code{\link{possibly_date}}, \code{\link{guessDateFormat}}
}
