% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2xlsx.R
\name{json2xlsx}
\alias{json2xlsx}
\title{Write a XLSX file from a JSON string}
\usage{
json2xlsx(json1, json2 = "{}", json3 = "{}", outfile, overwrite = FALSE,
  dll = FALSE)
}
\arguments{
\item{json1}{JSON string for the cells}

\item{json2}{JSON string for the images}

\item{json3}{JSON string for the passwords}

\item{outfile}{name of output file}

\item{overwrite}{logical, whether to overwrite `outfile` if it exists}
}
\value{
No returned value.
}
\description{
Write a XLSX file from a JSON string
}
\note{
This function is rather for internal purpose. The user should use
\code{\link{hwriteXLSX}}.
}
\examples{
colorMatrix <- matrix(grDevices::colours(), ncol=9)
f <- function(i, j){
  cell(j, i, value = colorMatrix[i,j], color = colorMatrix[i,j])
}
sheet <- list(Colors = do.call(Vectorize(f), expand.grid(i=1:73, j=1:9)))
json <- jsonlite::toJSON(sheet, null="null", auto_unbox = TRUE)
\donttest{json2xlsx(json, outfile="colors.xlsx")}
}
