% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{possibly_date}
\alias{possibly_date}
\title{Check whether conversion to a date is possible}
\usage{
possibly_date(x)
}
\arguments{
\item{x}{vector or factor}
}
\value{
A logical value, indicating whether \code{\link[base]{as.Date}} can be faithfully applied to \code{x}.
}
\description{
Check whether a vector can be converted to a date.
}
\examples{
possibly_date("2017-03-04")
possibly_date("2017-13-04")
}
\seealso{
\code{\link{is_date}}, \code{\link{guessDateFormat}}
}
